Moving files to correct location and app name
