+++ components/google_merchant_center/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pipedream/google_merchant_center",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "Pipedream Google Merchant Center Components",
   "main": "google_merchant_center.app.mjs",
   "keywords": [
@@ -11,5 +11,8 @@
   "author": "Pipedream <support@pipedream.com> (https://pipedream.com/)",
   "publishConfig": {
     "access": "public"
+  },
+  "dependencies": {
+    "@pipedream/platform": "^1.5.1"
   }
 }
